babygun/requirements.txt
absl-py==0.12.0
aiofiles==0.5.0
aniso8601==7.0.0
astor==0.8.1
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
dataclasses==0.6
dlib-bin==19.21.0
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
flatbuffers==1.12
Flask
gast==0.2.2
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.2.5
imutils==0.5.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
orjson==3.5.1
Pillow==8.2.0
promise==2.3
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
scipy==1.5.4
six==1.15.0
starlette==0.13.6
tensorboard==1.15.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.14.0
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yadisk==1.2.14
zipp==3.4.1
watchdog==2.1.2